Move UnpackedSequence to util. Remove files and functions in nn that are not ready for general use.
